upgrade library

+++ sapi/quickstart/linux/debian-init.sh
@@ -20,6 +20,7 @@ while [ $# -gt 0 ]; do
   shift $(($# > 0 ? 1 : 0))
 done
 
+
 case "$MIRROR" in
 china | ustc | tuna)
   OS_ID=$(cat /etc/os-release | grep '^ID=' | awk -F '=' '{print $2}')
@@ -27,16 +28,19 @@ china | ustc | tuna)
   case $OS_ID in
   debian)
     case $VERSION_ID in
-    11)
-      test -f /etc/apt/sources.list.save || cp /etc/apt/sources.list /etc/apt/sources.list.save
-      sed -i "s@deb.debian.org@mirrors.ustc.edu.cn@g" /etc/apt/sources.list
-      sed -i "s@security.debian.org@mirrors.ustc.edu.cn@g" /etc/apt/sources.list
-      test "$MIRROR" = "tuna" && sed -i "s@mirrors.ustc.edu.cn@mirrors.tuna.tsinghua.edu.cn@g" /etc/apt/sources.list
-      ;;
-    12)
-      test -f /etc/apt//etc/apt/sources.list.d/debian.sources.save || cp /etc/apt/sources.list.d/debian.sources /etc/apt/sources.list.d/debian.sources.save
-      sed -i 's/deb.debian.org/mirrors.ustc.edu.cn/g' /etc/apt/sources.list.d/debian.sources
-      test "$MIRROR" = "tuna" && sed -i "s@mirrors.ustc.edu.cn@mirrors.tuna.tsinghua.edu.cn@g" /etc/apt/sources.list.d/debian.sources
+    11 | 12 )
+      # 容器内和容器外 镜像源配置不一样
+      if [ -f /.dockerenv ]; then
+        test -f /etc/apt//etc/apt/sources.list.d/debian.sources.save || cp /etc/apt/sources.list.d/debian.sources /etc/apt/sources.list.d/debian.sources.save
+        sed -i 's/deb.debian.org/mirrors.ustc.edu.cn/g' /etc/apt/sources.list.d/debian.sources
+        sed -i 's/security.debian.org/mirrors.ustc.edu.cn/g' /etc/apt/sources.list.d/debian.sources
+        test "$MIRROR" = "tuna" && sed -i "s@mirrors.ustc.edu.cn@mirrors.tuna.tsinghua.edu.cn@g" /etc/apt/sources.list.d/debian.sources
+      else
+        test -f /etc/apt/sources.list.save || cp /etc/apt/sources.list /etc/apt/sources.list.save
+        sed -i "s@deb.debian.org@mirrors.ustc.edu.cn@g" /etc/apt/sources.list
+        sed -i "s@security.debian.org@mirrors.ustc.edu.cn@g" /etc/apt/sources.list
+        test "$MIRROR" = "tuna" && sed -i "s@mirrors.ustc.edu.cn@mirrors.tuna.tsinghua.edu.cn@g" /etc/apt/sources.list
+      fi
       ;;
     *)
       echo 'no match debian os version' . $VERSION_ID
